Add support to output accelerometer data in .npy format
Previous npy writer output uncalibrated data extracted directly from the accelerometer.
This is now changed to output calibrated data.
Bugs such as missing arguments to the npy writer are now fixed.
Note that the uncalibrated raw accelerometer data are integer values, whereas the calibrated values are of type double.
The time is stored in Unix nanoseconds and is of type long.