fix FFT and entropy normalization
To normalize in getFFTpower we need to divide by n^2 instead of n, as it
is the FFT coefs that get divided by n, then squared. This wouldn't be
an issue except when working with data with varying sampling rate, in
which case the features wouldn't be comparable.

The spectral entropy normalization is also fixed to use the whole
signal length.