auto download builtin ML models if not found
Download the builtin ML models during runtime if the model files are
not found. `--activityModel` flag is now to be interpreted as the builtin
model type to use (either `doherty2018` or `willetts2018`) rather than
the path to the model file. The code will then link to the corresponding
file via the dictionary `MODELS` in `accClassification.py`.

This change makes it easier to update our builtin models in the future
without having to notify the users to manually redownload the ML
models, and also `--activityModel` wouldn't need to change to the
updated path/filename. Instead, it will be on us to simply update
`MODELS` in `accClassification.py` to the new models.

Users can still use a custom activity model by supplying the path in
`--activityModelPath`. If supplied, then `--activityModel` is ignored.