native support for DST correction of CWA timestamp
Add native support to correct the CWA timestamps for potential
daylight saving time (DST) crossovers. To do this it keeps track of the
recording start time and its DST status, then for any subsequent
timestamp check if DST status has changed. If so, perform the
correction accordingly. For all this a timezone needs to be provided
(otherwise resolving DST wouldnt be possible) which defaults to
Europe/London.

All output files are now timezoned to UTC, although in the future it is
maybe preferable to convert it to the user-specified timezone, e.g.
Europe/London.