change timezone of output files to user-specified
The timezone of the output files are changed from UTC to the
user-specified timezone e.g. Europe/London. This requires a change
in the timeSeries.csv format: previously it had a column name that
detailed the start/end time and period to avoid having a column of
timestamps, but now it needs to have this because this time the
timestamps are timezone-aware and offsets may change in between due
to DST.

Two included functions will be generally useful to parse timestamps with
the outputted format:
accUtils.date_parser converts the formatted string to datetime
accUtils.strftime converts datetime to formatted string