handle potential nan/inf in correlation features
Handle nan/inf values due to precision error and/or degenerate
random variables:
1. Make sure that the variance is nonnegative before taking the sqrt.
2. If the returned correlation is undefined/nan (due to an axis being
all constant), we set the value to 0.0. That is, we shall assume the axes
are independent at rest. This is only done in Features.java and not in
the backend function AccStats.correlation() on the off chance that
someone use this function for something else that's not our signals.