improv nonwear code & summary more consistent
Improve code for nonwear detection. Also change the output wear time
stats to be consistent with the stats outputted for other labels. For
example, 'wearTime-mon(hrs)' was not very useful as it merges all Monday
data. Now 'wear-mon-avg' is outputted instead which is more consistent
with the other stuff ('sleep-mon-avg', etc.).